Testing modification

+++ Jenkinsfile
@@ -10,6 +10,7 @@ pipeline {
             app_version = 'v1'
             rollback = 'false'
             testPassed = false
+            requirementsInstalled = false
     }
     stages{
         stage('Testing'){
@@ -19,23 +20,26 @@ pipeline {
                 dir('frontend') {
                     sh "ls -al"
                     script{
-                        try{
-                            sh "sudo npm uninstall -g angular-cli @angular/cli"
-                            sh "sudo npm cache clean --force"
-                            sh "sudo npm install -g @angular/cli@8.3.25"
-                            sh "sudo npm install --save-dev @angular/cli@8.3.25" //Updates local version?
-                            sh "sudo npm install"
-                            sh "sudo npm i karma-cli"
-                            sh "rm -rf package-lock.json"
-                            // sh "if npm version > 5.0 delete package-lock.json"
-                            sh "sudo npm install karma-junit-reporter --save-dev"
-                            // sh "karma start -d"
-                            sh "sudo npm i -D puppeteer karma-chrome-launcher"
-                            sh 'ng test --watch false'
-                            // sh "sudo npm update"
-                        }catch(err){
-                            testPassed = false
-                        }
+                        if(requirementsInstalled == false){
+                            try{
+                                sh "sudo npm uninstall -g angular-cli @angular/cli"
+                                sh "sudo npm cache clean --force"
+                                sh "sudo npm install -g @angular/cli@8.3.25"
+                                sh "sudo npm install --save-dev @angular/cli@8.3.25" //Updates local version?
+                                sh "sudo npm install"
+                                sh "sudo npm i karma-cli"
+                                sh "rm -rf package-lock.json"
+                                // sh "if npm version > 5.0 delete package-lock.json"
+                                sh "sudo npm install karma-junit-reporter --save-dev"
+                                // sh "karma start -d"
+                                sh "sudo npm i -D puppeteer karma-chrome-launcher"
+                                sh 'ng test --watch false'
+                                // sh "sudo npm update"
+                            }catch(err){
+                                testPassed = false
+                            }
+                            requirementsInstalled == true
+                        }  
                     }
                 }
                 script{
